Rename playlists test to urlParams test
